requirements.txt
# Dependencies
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.2
python-dotenv==0.19.2

# Runtime
gunicorn==20.1.0
honcho>=1.0.1

# Code quality
pylint==2.9.3
flake8==3.7.9

# Testing
factory-boy==2.12.0
nose==1.3.7
pinocchio==0.4.2
httpie==2.3.0

# Code coverage
coverage==4.5.4
codecov==2.1.10
